requirements.txt
Django>=4.2.0
# gunicorn>=20.1.0
# psycopg2-binary
# django-redis>=5.2.0
# celery>=5.2.7 
django-widget-tweaks==1.5.0
pytz
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0


langchain-chroma==0.2.2
langchain-core>=0.1.15
langchain-openai>=0.0.5
langchain-community>=0.0.13
langgraph>=0.0.15
langgraph-cli[inmem]
langchain-graph-retriever[chroma]
chromadb>=0.4.22

django-tailwind>=3.8.0
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
python-docx>=1.0.0
html2text>=2020.1.16
docx2pdf>=0.1.8
unidecode>=1.3.8

